LayoutTests/fast/forms/datetimelocal/datetimelocal-editable-components/space-does-not-scroll-page.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/common.js"></script>
<script src="../../../../resources/js-test.js"></script>
<script src="../../../../resources/ui-helper.js"></script>
</head>
<body style="height: 3000px; overflow: scroll;">

<input id="input" type="datetime-local" value="2020-09-16T10:45"/>

<script>
window.jsTestIsAsync = true;

addEventListener("load", async () => {
    description("Tests that the page is not scrolled when pressing space to activate the date picker.");

    await UIHelper.ensurePresentationUpdate();
    // Tab to focus the first subfield of the date input.
    UIHelper.keyDown("\t", ["altKey"]);
    // Press space to open the picker.
    UIHelper.keyDown(" ");

    await UIHelper.ensurePresentationUpdate();
    showingPicker = await UIHelper.isShowingDateTimePicker();
    shouldBeTrue("showingPicker");

    // Pressing the space key should not have scrolled the page.
    shouldBeTrue("window.scrollY === 0");
    shouldBeTrue("window.scrollX === 0");

    finishJSTest();
});
</script>
</body>
</html>
