LayoutTests/fast/forms/datetimelocal/datetimelocal-editable-components/space-does-not-scroll-page-expected.txt

Tests that the page is not scrolled when pressing space to activate the date picker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS showingPicker is true
PASS window.scrollY === 0 is true
PASS window.scrollX === 0 is true
PASS successfullyParsed is true

TEST COMPLETE
